First functional version of Delete query composer and conductor in TableQuery module. - Composer builds a single query per table, filtering each on all fields given in delete filters array. - Conductor executes every query given to it in a flat MultiQueryWrapper, as outputted by Composer.
For a "delete by ID" type query, this will need to be used with only the primary key field data passed as filters. Don't just fetch all data for a record and pass it straight through.
The reason we consider all fields in the delete request is so that records can be bulk deleted by filtering on fields other than ID.